==> online-sample/Intent-classification/1. Data Preprocessing.ipynb <==
### 1. Training & Testing json files present in data/train & data/test folder

import os
import json
import pandas as pd

path = './data/train'
train_json_files = [f for f in os.listdir(path) if f.endswith('.json')]
train_json_files

### 2. Loading and parsing the training data

data_classify = [ [1,0,0,0],[0,1,0,0],[0,0,1,0],[0,0,0,1] ]
count = 0

def parsing_logic(data,train_data_name,count):
    final_dataset = []
    for sentence in data[train_data_name]:
        words = sentence['data']
        temp = ''
        for word in words:        
            temp = temp + word['text']
        final_dataset.append([temp.rstrip(), data_classify[count][0],data_classify[count][1],
                            data_classify[count][2],data_classify[count][3]]) # rstrip() -> Stripping next line
    return final_dataset

final_datasets = []    
for train_json_file in train_json_files:
    try:
        with open('./data/train/'+train_json_file, encoding="utf8") as data_file:    
            data = json.load(data_file)
    except:
        with open('./data/train/'+train_json_file, encoding="Latin-1") as data_file:    
            data = json.load(data_file)
    start_index = 8
    end_index = train_json_file.find('_full')
    train_data_name = train_json_file[start_index:end_index] 
    
    print(train_json_file)
    print(len(data[train_data_name]))
    final_dataset = parsing_logic(data,train_data_name,count)
    count = count + 1
    print(len(final_dataset))
    final_datasets = final_datasets + final_dataset
    print('----------------------------------------------------')

final_datasets

### 3. Loading the data in Pandas Dataframe

df = pd.DataFrame(final_datasets,columns=['sentence','BookRestaurant','GetWeather','PlayMusic','RateBook'])

df.head()

len(df)

df.to_csv('./data/train/train.csv', index = False,  encoding='utf-8')
